+++ lie_nn/_src/rep.py
@@ -3,7 +3,7 @@ from typing import Iterator, Optional, Tuple, Type
 import numpy as np
 import scipy.linalg
 
-from .util import check_algebra_vs_generators, direct_sum, infer_algebra_from_generators
+from .util import direct_sum, infer_algebra_from_generators
 
 
 class Rep:
@@ -76,6 +76,8 @@ class Rep:
         Returns:
             ``(dim, dim)`` array
         """
+        # TODO: now that we integrate the Sn group, that is non abelian.
+        # there is no more rule to how to parameterize the finite part of the represention.
         output = scipy.linalg.expm(
             np.einsum("a,aij->ij", continuous_params, self.continuous_generators())
         )
@@ -91,11 +93,6 @@ class Rep:
             self.dim == 1
             and np.all(self.continuous_generators() == 0.0)
             and np.all(self.discrete_generators() == 1.0)
-        )
-
-    def check_algebra_vs_generators(rep: "Rep", rtol=1e-10, atol=1e-10):
-        check_algebra_vs_generators(
-            rep.algebra(), rep.continuous_generators(), rtol=rtol, atol=atol, assert_=True
         )
 
 
@@ -190,7 +187,9 @@ class MulRep(Rep):
     mul: int
     rep: Rep
 
-    def __init__(self, mul: int, rep: Rep):
+    def __init__(self, mul: int, rep: Rep, *, force=False):
+        if not force:
+            raise RuntimeError("Use lie_nn.multiply instead")
         self.mul = mul
         self.rep = rep
 
@@ -236,7 +235,9 @@ class SumRep(Rep):
     """
     reps: Tuple[Rep, ...]
 
-    def __init__(self, reps: Tuple[Rep, ...]):
+    def __init__(self, reps: Tuple[Rep, ...], *, force=False):
+        if not force:
+            raise RuntimeError("Use lie_nn.direct_sum instead")
         assert len(reps) >= 1
         self.reps = tuple(reps)
 
@@ -286,7 +287,9 @@ class QRep(Rep):
     rep: Rep
     Q: np.ndarray
 
-    def __init__(self, Q: np.ndarray, rep: Rep):
+    def __init__(self, Q: np.ndarray, rep: Rep, *, force=False):
+        if not force:
+            raise RuntimeError("Use lie_nn.change_basis instead")
         self.rep = rep
         self.Q = Q
 
@@ -308,58 +311,3 @@ class QRep(Rep):
 
     def __repr__(self) -> str:
         return f"Q({self.rep})Q^{{-1}}"
-
-
-def check_representation_triplet(rep1: Rep, rep2: Rep, rep3: Rep, rtol=1e-10, atol=1e-10):
-    assert np.allclose(rep1.algebra(), rep2.algebra(), rtol=rtol, atol=atol)
-    assert np.allclose(rep1.algebra(), rep3.algebra(), rtol=rtol, atol=atol)
-
-    rep1.check_algebra_vs_generators(rtol=rtol, atol=atol)
-    rep2.check_algebra_vs_generators(rtol=rtol, atol=atol)
-    rep3.check_algebra_vs_generators(rtol=rtol, atol=atol)
-
-    X1 = rep1.continuous_generators()  # (lie_group_dimension, rep1.dim, rep1.dim)
-    X2 = rep2.continuous_generators()  # (lie_group_dimension, rep2.dim, rep2.dim)
-    X3 = rep3.continuous_generators()  # (lie_group_dimension, rep3.dim, rep3.dim)
-    assert X1.shape[0] == X2.shape[0] == X3.shape[0]
-
-    from .clebsch_gordan import clebsch_gordan
-
-    cg = clebsch_gordan(rep1, rep2, rep3)
-    assert cg.ndim == 1 + 3, (rep1, rep2, rep3, cg.shape)
-    assert cg.shape == (cg.shape[0], rep1.dim, rep2.dim, rep3.dim)
-
-    # Orthogonality
-    # left_side = np.einsum('zijk,wijl->zkwl', cg, np.conj(cg))
-    # right_side = np.eye(cg.shape[0] * rep3.dim)
-    # .reshape((cg.shape[0], rep3.dim, cg.shape[0], rep3.dim))
-    # np.testing.assert_allclose(left_side, right_side, rtol=rtol, atol=atol)
-
-    # if rep3 in rep1 * rep2:
-    #     assert cg.shape[0] > 0
-    # else:
-    #     assert cg.shape[0] == 0
-
-    left_side = np.einsum("zijk,dlk->zdijl", cg, X3)
-    right_side = np.einsum("dil,zijk->zdljk", X1, cg) + np.einsum("djl,zijk->zdilk", X2, cg)
-
-    for solution in range(cg.shape[0]):
-        for i in range(X1.shape[0]):
-            if not np.allclose(
-                left_side[solution][i], right_side[solution][i], rtol=rtol, atol=atol
-            ):
-                np.set_printoptions(precision=3, suppress=True)
-                print(rep1, rep2, rep3)
-                print('Left side: einsum("zijk,dlk->zdijl", cg, X3)')
-                print(left_side[solution][i])
-                print(
-                    "Right side: "
-                    'einsum("dil,zijk->zdljk", X1, cg) + einsum("djl,zijk->zdilk", X2, cg)'
-                )
-                print(right_side[solution][i])
-                np.set_printoptions(precision=8, suppress=False)
-                raise AssertionError(
-                    f"Solution {solution}/{cg.shape[0]} for {rep1} * {rep2} = {rep3} "
-                    "is not correct."
-                    f"Clebsch-Gordan coefficient is not correct for Lie algebra generator {i}."
-                )
