+++ lie_nn/test.py
@@ -1,0 +1,3 @@
+from ._src.tests import check_representation_triplet
+
+__all__ = ["check_representation_triplet"]
